Clean install directory before installing
OpenSSL 1.1.x is already installed on the GitHub actions Windows VMs.
When running `nmake install`, our files get merged into the same
directory.  This causes the `7z a -tzip` command to pick up both the
files we installed and whatever was already there.  Let's ensure we
get a clear archive!