Fix CMake test on 64-bit machines
On 64-bit machines, our `ctest` run would fail for OpenSSL 1.1.1s
because something else in PATH provided `libssl-1_1-x64.dll` as well.
It's not clear what was conflicting on the GitHub runner VMs, but I
reproduced this on my machine with Git for Windows.

We didn't encounter this for OpenSSL 3.x, nor for OpenSSL 1.1.1s
32-bit binaries, but technically it could happen for any other OpenSSL
version and/or architecture depending on what software is installed on
the machine.

Work around the issue by forcing the path to our binaries at the front
of `PATH` while running tests.

Note that this wouldn't be an acceptable solution when installing the
binaries for a real program.  If you were to prepare a real
application and install it, you should probably copy the DLLs and
install them in the same folder as your executable.