Copy OpenSSL dynamic-link libraries to the output folder
I finally tracked down the issue for OpenSSL 1.1.x on GitHub Windows
runners.  The root cause is that the runners have `libssl-1_1-x64.dll`
in `C:\Windows\system32`.  Given the documented Dynamic-link library
search order (see Windows documetnation), this file is found before
`PATH` is searched.  The ONLY path that is searched before
`C:\Windows\system32` is the directory containing the application
executable.  This means the ONLY way to resolve our OpenSSL binaries
on this machine is to copy them in the same folder as the executable
and we can never correctly resolve with PATH, no matter the order in
which we add our OpenSSL binary directory in there.

This is rather unfortunate because CMake's FindOpenSSL module doesn't
expose any helpers to achieve this and we have to compute the path to
the DLLs relative to `${OPENSSL_CRYPTO_LIBRARY}` or
`${OPENSSL_SSL_LIBRARY}.