Add static build variant
Although the DLL (shared library) build is the default option when
building OpenSSL from source, it's rather impractical on Windows, but
it requires you to bundle `libcrypto*.dll` and `libssl*.dll` with your
app.

In addition, some existing CMake-based projects have issues where they
build correctly, but don't copy the OpenSSL DLLs into the output
folder (normal, as the `FindOpenSSL` module in CMake doesn't explain
that this is required and doesn't facilitate it either).  This causes
the application to crash on boot due to the missing DLLs.

The whole intent of installing OpenSSL as a shared library is to ship
security fixes without recompiling all applications on the system, but
this doesn't work on Windows because OpenSSL is not installed
centrally.  If you're bundling the DLLs directly alongside your
application (which you should, see the README), then you'll have to
publish a bugfix update to your application anyways and it's not too
much trouble to recompile it anyways.  Let's offer the static build
for people who prefer the static linking option.